Move inner classes to packages
